+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Color semantic tokens (tokens library v1.1.0) (Orange-OpenSource/ouds-ios#664)
 - Space semantic tokens (tokens library v1.1.0) (Orange-OpenSource/ouds-ios#664)
 - Border semantic tokens (tokens library v1.1.0) (Orange-OpenSource/ouds-ios#664)
 - Size semantic tokens (tokens library v1.1.0) (Orange-OpenSource/ouds-ios#664)